Renamed Income to IncomeModel
